@@ -7,13 +7,17 @@ exports.getUsers = (req, res) => {
     .catch((err) => res.status(400).json(err));
 };
 
-exports.addUsers = (req, res) => {
-  // creating a new instance of User using username
-  const newUser = new User(req.body);
+exports.addUsers = async (req, res) => {
+  const { username } = req.body;
 
-  // saving newUSer to mongoDB
-  newUser
-    .save()
-    .then(() => res.json('added User'))
-    .catch((err) => res.status(400).json(err));
+  // creating a new instance of User using username
+  const newUser = new User({ username });
+
+  try {
+    // saving newUSer to mongoDB
+    await newUser.save();
+    res.status(201).json('added user');
+  } catch (error) {
+    res.status(400).json({ message: error.message });
+  }
 };
