@@ -6,7 +6,7 @@ const workoutsController = require('../controller/workouts');
 router.get('/', workoutsController.getWorkout);
 router.post('/add', workoutsController.addWorkout);
 router.get('/:id', workoutsController.editWorkout);
-router.post('/update/:id', workoutsController.updateWorkout);
+router.patch('/update/:id', workoutsController.updateWorkout);
 router.delete('/:id', workoutsController.deleteWorkout);
 
 module.exports = router;
