@@ -3,11 +3,11 @@ require('dotenv').config();
 const express = require('express');
 const morgan = require('morgan');
 const connectDB = require('./config/db');
+const cors = require('cors');
 
 const app = express();
 
 // Middleware
-const cors = require('cors');
 app.use(cors());
 app.use(express.json()); // Parsing JSON
 app.use(morgan('dev')); // http request logger
